// src/app/commons/models/intermediaryChannel.ts
import {CanalDetail} from './canalDetail';

export class IntermediaryChannel {
  id?: string;
  index?: number;
  name: any;
  serveOtherSegments: any;
  estimatedGrossMargin: any;
  averageSaleCanal: any;
  signContracts: any;
  assignsAccountsTerritories: any;
  canalDetail: CanalDetail[];
  smallView: boolean;
  idCanalSegment: string;
}